// Programacion3/TP Integrador/Entrega/frontend/final-prog3/src/types/product.ts
import type {ICategory} from "./category"
export interface Product{
    id: number;
    eliminado: boolean;
    createdAt: string;
    nombre: string;
    precio: number;
    descripcion: string;
    stock: number;
    imagen: string;
    disponible: boolean;
    categoria: ICategory;
}

export interface CartItem{
    product:Product;
    quantity: number;
}